foundation(connections):
established a connection between the go, html, css, js files and the databse. usage of a single page application plan for future project (real time forum), attempting to add a dockerfile from now on, needs testing. a readme start for the plan and envision of the project. error handling.

the second part, we established a initial functionality, which is for the login and the signup and the logout. the handling of cookies for the user, knowing which the user is logged on or logged out and displaying a different page depending on it.